* commit 'd9e6b55d9f6c22cfd31fbfe3b6979d5d25025512':
  Adding the delete work item code on regression only
  Refactored the code to delete work items

+++ src/main/java/com/apas/Listeners/SuiteListener.java
@@ -65,8 +65,11 @@ public class SuiteListener extends TestBase implements ITestListener {
 			TestBase.loadPropertyFiles();
 			extent = new ExtentManager().getInstance(context.getSuite().getName());
 			if (flagToUpdateJira && testCycle != null) {
-				salesforceAPI.deleteWorkItemsBasedOnAge(1);
 				JiraAdaptavistStatusUpdate.retrieveJiraTestCases();
+				if (CONFIG.getProperty("deleteWorkItemsFlag").equals("true")){
+					System.out.println("Deleting the work items for the age greater than 1");
+					salesforceAPI.deleteWorkItemsBasedOnAge(1);
+				}
 			}
 
 			//This will move old report to archive folder
